+++ src/backend/resolvers/mutations/payerProject/index.js
@@ -1,9 +1,11 @@
 const updatePayerProjectPtps = require('./updatePtps')
 const removePayerProjectPtps = require('./removePtps')
 const transferPayerProjectPtps = require('./transferPtps')
+const createPayerProject = require('./create')
 
 module.exports = {
   updatePayerProjectPtps,
   removePayerProjectPtps,
   transferPayerProjectPtps,
+  createPayerProject,
 }