+++ src/backend/typeDefs/mutations/payerProject/create.js
@@ -1,0 +1,16 @@
+const { gql } = require('apollo-server-express')
+
+const createPayerProject = gql`
+  input CreatePayerProjectInput {
+    name: String!
+  }
+
+  type CreatePayerProjectPayload {
+    _id: ID!
+    name: String
+    orgTpIds: [ID]
+    extraOrgTpIds: [ID]
+  }
+`
+
+module.exports = createPayerProject