+++ src/backend/typeDefs/mutations/payerProject/index.js
@@ -1,9 +1,11 @@
 const updatePayerProjectPtpsTypeDefs = require('./updatePtps')
 const removePayerProjectPtpsTypeDefs = require('./removePtps')
 const transferPayerProjectPtpsTypeDefs = require('./transferPtps')
+const createPayerProjectTypeDefs = require('./create')
 
 module.exports = [
   updatePayerProjectPtpsTypeDefs,
   removePayerProjectPtpsTypeDefs,
   transferPayerProjectPtpsTypeDefs,
+  createPayerProjectTypeDefs,
 ]