+++ src/frontend/api/mutations/payerProjects.js
@@ -1,5 +1,16 @@
 import gql from 'graphql-tag'
 
+export const CREATE_PAYER_PROJECT = gql`
+  mutation CreatePayerProject($input: CreatePayerProjectInput!) {
+    createPayerProject(input: $input) {
+      _id
+      name
+      orgTpIds
+      extraOrgTpIds
+    }
+  }
+`
+
 export const UPDATE_PAYER_PROJECT_PTPS = gql`
   mutation UpdatePayerProjectPtps($input: UpdatePayerProjectPtpsInput!) {
     updatePayerProjectPtps(input: $input)